Update README.md
Caesar Cipher tool                                                                                                              This enables individuals to safely encrypt or decrypt texts with particular shift values. For secure insertion of sensitive input such as the message and shift amount, getpass should be employed. Moreover, this script has facilities for analyzing message composition by showing how many letters, digits and special characters are there. This makes it easier for users to know what option they would choose from when they are prompted whether to encrypt, decrypt or exit from the system. In general, the tool is built on cryptographic functions but its interaction is user friendly making it suitable for educational purposes or basic encryption needs in a command-line environment.